liquidex: (de)serialize blinders correctly
Asset and value blinders are serialized in reversed when displayed
in hex. This was not happening, and thus the test vector from
the python version of LiquiDEX was failing commitment verification.

After this commit, blinders are (de)serialized (from)to hex
correctly, and we can verify commitments on the test vector.