wollet: fix last_index update check
First rename last_index to last_unused.
Storing the last used index is incorrect, because when we first
create a wallet, the last used index is "-1", but we can't express
it with a u32.
So we store instead the last _unused_ index, then in sync we
compare the last unused index from the store with the last unused
index from the latest sync cycle.