jade/wollet: singlesig issueasset use change and external
Before this commit jade would display 4 outputs for an issuance:
issuance, reissuance token, lbtc change and fee. This happens because all
are change outputs for its so instead of hiding them all decides to show them
all.
With this change the descriptor used in the test is multipath so the wallet
can generate both external and internal address, moreover issueasset by default
creates recipient for the issued asset and the reissuance token to external
address.