cli: fix logging
This remove the init_logging fn because having logging initialization in a
sub-scope make it not working probably because something is dropped.

Instead of relying on the sync once mechanism we simply avoid throwing the
error when the initialization is done twice in tests.

Moreover, tests prints on stderr by default instead of debug.log