jade: unlock jade before doing actual calls
If the Jade is locked it can't do most calls.

In practice, Jade might lock itself for inactivity or the user
can lock it manually.

We can either rely on the caller to catch the error that signals
that Jade is locked and unlock Jade in that case, or we can
unlock Jade before any calls.

We opted for the latter, since it makes it easier for the library
users. Note that if Jade is already unlocked, we just need to
query Jade once, so this seems a reasonable trade off.