introduce uniffi bindings
This adds a new crate ks_bindings to the workspace containing the bindings
generator based on mozilla uniffi and a couple of methods to showcase it.

It's just a PoC but it should not influence the working of the other flows
other than compile times. To achieve so the binary to generate the bindings
is integrated as an hidden command in our cli.