wollet: add fixed addresses test
The previous commit changes the fixed addresses used in the test (from testnet
to regtest). That fixed address caught a bug in the process of updating
rust-elements to rust-bitcoin 0.31, see https://github.com/ElementsProject/rust-elements/pull/188

This introduced tests will check any possible singlesig variant descriptor against
a static address in the hope to prevent similar cases to happen in the future.