tests: do persist check twice
At the moment there is a bug, since we are persisting updates also during the
initial loading of the wallet. This cause to duplicate updates.
I thought this was a way to expose the issue but it's not the case, presumably
becuase our cache is not sensitive to duplicate updates.
Checking twice may discover regression in the future so I'll keep this.