wasm: always accept borrowed data unless explicit
In javascript it's not common to "move" data in a function, thus it's
unexpected when you can't use anymore a variable passed to a function.
For this reason always accept borrowed data in exposed function such that the
caller can continuing to use them after passing those to a function.

The only exception is when we return modified data and we don't want the caller
to think the passed variable is mutated, so he must use the returned value.