jade: Introduce Stream trait
To work in the browser we need a wasm implementation implementing our trait,
we can't impl AsyncRead (on which AsyncReadExt is built upon) because we have
the poll fn which isn't async, thus we have to introduce our trait and we
must implement it for the tcp stream.

This also remove tokio_serial for now which doesn't work in the browser.