jade: make auth jade driven
the v2 version of the pin protocol requires a different flow between jade
and requests made to the pin server. To be compatible with the old method, v2
and newer methods the code handling the auth is made generic, following which
calls and which requests the host has to do for the Jade.