wollet: get address status
we currently don't support subscribing to addresses but we only support full
scan. Changing or adding support properly is a big job and may require using
threads which come at costs.
In some case you want to monitor an address (eg donation page) and react.
With this you can poll on the address and do a full scan if change on the
address is detected, avoiding to poll on a full scan instead.