wollet: reissue_asset: allow to pass issuance_tx
`reissue_asset` relies on the fact that the `Wollet` is aware of
the issuance transaction. However `Wollet` only knows transactions
involving scripts derived from its descriptor.
In case the reissuance token is sent to another wallet, we need
to provide the issuance information.
We do so by adding an optional parameter for the issuance
transaction.

In this way `Wollet` data remains obtainable from the
descriptor + blockchain source only.

We don't actually need the whole issuance transaction
(we only need token and entropy), but this makes it harder to
misuse by callers (e.g. serialize incorrectly entropy).