wasm: use singlesig_desc
To use the standard function to create singlesig descriptors we had an
issue in async world since the standard Signer trait is syncronous.
To overcome this, we create a fake signer struct with all the possible
singlesig derivation and pass that instead.
This is hacky, but at least the delicate part of creating descriptors
is not duplicated.