wollet: fix only tip for wasm client
without this the generated Update returns always false for `only_tip`
becuase timestamps are always included. This logic instead insert those
only if missing in the db and avoids the web app to save a persisted
update at every scan.

Unfortunately it is missing test because it needs the waterfall endpoint
to be included in tests.