wollet: derive Address on demand
deriving the address (especially the blinded one, but also the unblinded)
is an expensive operation and we don't want to pay the price for every
wallet methods (such as transactions()) if the caller doesn't need it,
like in most cases. So we keep the definite descriptor in WalletTxOut
and derive the address on the fly if the WalletTxOut::address() method
is called.
Bench of the call to transactions() with 63 txs went from about 18ms to
400us.