tiny_jrpc: remove reqwest dev dep
The dep is causing issues in the nix env where the builder fails
presumably trying to get the TLS stack. Even removing the TLS features
from the dep the error persist, presumably because reqwest is used in
other crates and the feature is loaded anyway.
The library is replaced by using TCPStream directly, vibe coded with
Claudio.