wollet: tx builder: liquidex make: create blind asset proof
Under the non-LiquiDEX flow, the blind asset proof is created by
rust-elements' blinding functions. But with LiquiDEX that is not
the case, so create one here manually.

Note that this is not strictly necessary as one could verify that
the pset asset commitment matches the value computed from the
(explicit) asset and the asset blinding factor. Anyway this makes
it easier for programs that need to analyze PSETs, as they don't
have to handle LiquiDEX maker PSETs in a special way.