docs: remove rust snippets
Rust snippets can be anywhere in the code.
For instance now are in lwk_wollet/tests/e2e.rs.
While this approach is more disperse, it allows us to leverage
the existing testing infrastructure, which sets up regtest env,
electrum/esplora/waterfalls servers, interacts with emulators
for Jade and Ledger, and more.
This results in less code, and less tests to set up and run.