common: pset_balance: use blind proofs
If we require the rangeproofs, then "wollet"s need to keep the
rangeproofs in order to put them into the PSET inputs. This is not
ideal since rangeproofs are quite large.

Using the blind proofs is an equivalent alternative to prove that
asset and value commit to the respective commitments, while
allowing wollets to avoid storing the output rangeproofs.

However for backward compatbility, if there are no blind proofs,
we still try to unblind the previous output (using the rangeproof)
before failing.