payment_instr: remove silentpayments dep for now
while it's good to have a proper silentpayment address parsing, since we
don't support it fully it doesn't add much and since the crate use a
previous version of secp, it brings in another copy of secp256k1, and we
don't want that. We can revisit if the crate upgrade.