boltz: fix restore froms swap list test
A swap can be restored and completed if the tx has been made before being
interrupted.
Otherwise we haven't the amount needed to complete the swap. A swap
that is created but not payed is not necessarily to be completed or
refunded because no user fund was used in the first place.