boltz: avoid restoring a swap in status swap.created
a swap in swap.created has just been created, no funds from the user has
been sent so there is no fund loss from the user. We simply avoid to try
to restore those, even because we haven't all the information to
complete those in the first place, the boltz api doesn't return the api
amount needed to be sent.